requirements.txt
# Python 3.10-3.11 recommended
numpy>=1.24
pandas>=1.5
scikit-learn>=1.3
plotly>=5.17
vectorbt>=0.24
PyYAML>=6.0
requests>=2.31
pytest>=7.4
